Fix incompatible dependency
react-file-reader-input have a dependency error and every installation of a library, the error comes. Initially, I used --force to install react-file-reader-input to get rid of that, but now the issue is now solved.

The issue is solved by overriding the react-file-reader-input's properties, and I also updated the react version to the latest.

Tensorflow is installed after the dependency is solved. It will be used for emotion detection model.